# assets/openapi.yaml
openapi: 3.0.3

info:
  title: HTTB.dev API Docs
  version: unstable
  description: "HTTB is a modern HTTP bin. Test your HTTP client or proxy with various HTTP methods, status codes, and more."
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: MarvinJWendt
    url: https://github.com/MarvinJWendt

servers:
  - url: https://httb.dev
    description: Official HTTB server
  - url: http://localhost:{port}
    description: Local HTTB server
    variables:
      port:
        default: "8080"
        description: "Port of the local server"
  - url: "{custom}"
    description: Custom HTTB server
    variables:
      custom:
        default: "http://localhost:8080"
        description: "URL of the custom server"

paths:
  # --- Special Endpoints ---
  /ping:
    get:
      tags:
        - Special Endpoints
      summary: "Returns `pong`."
      description: "Very simple endpoint that always returns `pong`."
      parameters:
        - $ref: '#/components/parameters/FormatParam'
        - $ref: '#/components/parameters/DelayParam'
      responses:
        '200':
          description: "Returns `pong`."
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: "pong"
            text/plain:
              example: "pong"
              schema:
                type: string
                description: "Latency in milliseconds"

  # --- JSON Endpoints ---
  /json/random:
    get:
      tags:
        - JSON
      summary: Returns random JSON data.
      description: Returns JSON data.
      parameters:
        - $ref: '#/components/parameters/DelayParam'
      responses:
        '200':
          description: JSON data.
          content:
            application/json:
              schema:
                type: object

  /json/random/log:
    get:
      tags:
        - JSON
      summary: Returns random log data.
      description: Returns log data.
      parameters:
        - $ref: '#/components/parameters/DelayParam'
        - $ref: '#/components/parameters/LogLevels'
        - $ref: '#/components/parameters/LogLevelWeights'
        - name: count
          in: query
          description: "Number of log entries to return (min: 1; max: 10000)"
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 10
      responses:
        '200':
          description: Log data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Log'

  /json/random/address:
    get:
      tags:
        - JSON
      summary: Returns random address data.
      description: Returns address data.
      parameters:
        - $ref: '#/components/parameters/DelayParam'
      responses:
        '200':
          description: Address data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'

  /json/random/contact:
    get:
      tags:
        - JSON
      summary: Returns random contact data.
      description: Returns contact data.
      parameters:
        - $ref: '#/components/parameters/DelayParam'
      responses:
        '200':
          description: Contact data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'

  /json/random/user:
    get:
      tags:
        - JSON
      summary: Returns random user data.
      description: Returns user data.
      parameters:
        - $ref: '#/components/parameters/DelayParam'
      responses:
        '200':
          description: User data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # --- Streaming Endpoints ---
  /stream/json/user:
    get:
      tags:
        - JSON
        - Stream
      summary: "Streams JSON data."
      description: "Streams JSON data."
      parameters:
        - $ref: '#/components/parameters/DelayParam'
        - $ref: '#/components/parameters/StreamInterval'
      responses:
        '200':
          description: "JSON data."
          content:
            application/json:
              schema:
                type: object


  /stream/json/logs:
    get:
      tags:
        - JSON
        - Stream
      summary: "Streams JSON logs."
      description: "Streams JSON logs."
      parameters:
        - $ref: '#/components/parameters/DelayParam'
        - $ref: '#/components/parameters/LogLevels'
        - $ref: '#/components/parameters/LogLevelWeights'
        - $ref: '#/components/parameters/StreamInterval'
      responses:
        '200':
          description: "JSON logs."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Log'

components:
  schemas:
    Address:
      type: object
      properties:
        street:
          type: string
          example: "1234 Elm St."
          x-oapi-codegen-extra-tags:
            fake: "{street}"
        city:
          type: string
          example: "Springfield"
          x-oapi-codegen-extra-tags:
            fake: "{city}"
        zip:
          type: string
          example: "62701"
          x-oapi-codegen-extra-tags:
            fake: "{zip}"
    Contact:
      type: object
      properties:
        email:
          type: string
          format: email
          x-oapi-codegen-extra-tags:
            fake: "{email}"
        phone:
          type: string
          format: phone
          x-oapi-codegen-extra-tags:
            fake: "{phone}"
    User:
      type: object
      properties:
        firstName:
          type: string
          example: "John"
          x-oapi-codegen-extra-tags:
            fake: "{firstname}"
        lastName:
          type: string
          example: "Doe"
          x-oapi-codegen-extra-tags:
            fake: "{lastname}"
        age:
          type: integer
          minimum: 0
          maximum: 100
          example: 42
        gender:
          type: string
          enum:
            - male
            - female
            - other
          x-oapi-codegen-extra-tags:
            fake: "{randomstring:[male,female,other]}"
        contact:
          type: object
          $ref: '#/components/schemas/Contact'
        address:
          type: object
          $ref: '#/components/schemas/Address'
    Log:
      type: object
      properties:
        level:
          type: string
          example: "info"
          x-oapi-codegen-extra-tags:
            fake: "{weighted:[debug,info,warn,error],[1,5,3,2]}"
        message:
          type: string
          example: "We need to calculate the virtual PNG capacitor!"
          x-oapi-codegen-extra-tags:
            fake: "{hackerphrase}"
        timestamp:
          type: string
          format: date-time

  parameters:
    FormatParam:
      name: format
      in: query
      description: "Response format (default: `json`)"
      required: false
      schema:
        type: string
        enum: [ json, text ]
        default: json
    DelayParam:
      name: delay
      in: query
      description: "Delay in milliseconds before the response is sent (min: 0; max: 10000)"
      required: false
      schema:
        type: integer
        minimum: 0
        maximum: 10000
        default: 0
    LogLevels:
      name: logLevels
      in: query
      description: "Log levels to use (default: `debug,info,warn,error`)"
      schema:
        type: array
        items:
          type: string
        default: [ debug, info, warn, error ]
    LogLevelWeights:
        name: logLevelWeights
        in: query
        description: "Log level weights (default: `1,5,3,2`)"
        schema:
          type: array
          items:
            type: number
          default: [ 1, 5, 3, 2 ]
    StreamInterval:
        name: interval
        in: query
        description: "Interval in milliseconds between streamed responses (min: 0; max: 5000)"
        schema:
            type: integer
            minimum: 0
            maximum: 5000
            default: 250

tags:
  - name: Special Endpoints
    description: "Endpoints providing special functionality."
  - name: Return Request Data
    description: "Endpoints returning the received request data."
  - name: HTTP Methods
    description: "Endpoints to test different HTTP methods."
  - name: Status Codes
    description: "Endpoints to test various HTTP status codes."
  - name: JSON
    description: "Endpoints returning JSON data."
  - name: Stream
    description: "Endpoints returning streaming data."